Move pending_validations.rs from scheduled jobs to dht and run_holding_workflow() to workflows
