+++ crates/core/src/dht/mod.rs
@@ -3,5 +3,6 @@
 pub mod actions;
 pub mod dht_reducers;
 pub mod dht_store;
+pub mod pending_validations;
 
 mod dht_inner_reducers;
