+++ crates/core/src/scheduled_jobs/mod.rs
@@ -1,4 +1,3 @@
-pub mod pending_validations;
 pub mod state_dump;
 
 use crate::context::Context;
